// next-shell/src/types/term_form.types.ts
export interface TermFormData {
    supplier: string;
    supplierName: string;
    contactPerson: string;
    contactPersonName: string;
    contractNo: string;
    mfgPartNo: string;
    suppOrderCode: string;
    active: boolean;
    validFrom: string;
    validTo: string;
    purchaseTerm: string;
    purchaseTermLocation: string;
    purchaseTermLocationName: string;
    purchaseSubLocation: string;
    salesTerm: string;
    salesSubLocation: string;
    incoterm: string;
    prodCost: number;
    pkh: number;
    soc: number;
    currency: string;
    exRate: number;
    shipMode: string;
    dimUnit: string;
    length: number;
    width: number;
    height: number;
    weight: number;
    cWeight: number;
    freightType: string;
    freightRate: number;
    fr: number;
    insPercent: number;
    dutyPercent: number;
    excisePercent: number;
    miscTax: number;
    numInBuy: number;
    numInSale: number;
    markup: number;
    stockFeePercent: number;
    wireTT: number;
    customClear: number;
    scc: number;
    zoneRate: number;
    salesPerson: string;
    salesPersonName: string;
    sourcedBy: string;
    sourcedByName: string;
    remark: string;
    leadTime: string;
    moq: string;
    vendorBPA: boolean;
    purchaseUOM: string;
    salesUOM: string;
    spk: number;
    qoc: number;
    updatedBy: string;
    updatedDate: string;
    stockUOM: string;
}

export interface TermCalculationPayload {
    U_ProdCost: number;
    U_PKH: number;
    U_SOC: number;
    U_PurRate: number;
    U_OrderTerm: string;
    U_ShipModeNo: number;
    U_DimUnitNo: number;
    U_Length: number;
    U_Width: number;
    U_Height: number;
    U_Weight: number;
    U_FreightRate: number;
    U_FR: number;
    INS_Percent: number;
    U_ZoneRate: number;
    U_DT_Percent: number;
    U_ETPer: number;
    U_MiscTax: number;
    U_WTT: number;
    U_CC: number;
    U_ASP: number;
    U_STK_Percent: number;
    U_SPK: number;
    U_QOC: number;
    U_MK_Percent: number;
    NumInBuy: number;
    NumInSale: number;
}

export interface TermCalculationResponse {
    U_OP: number;
    U_OP_SUM: number;
    U_OP_THB: number;
    U_DimWeight: number;
    U_ShipWeightCal: number;
    U_INS: number;
    U_FRZONE: number;
    U_FreightQTEC: number;
    U_CIF: number;
    U_CIFZONE: number;
    U_ZoneRate: number;
    U_DT: number;
    U_DT_FR: number;
    U_DT_FRZONE: number;
    U_ET: number;
    U_MT: number;
    U_preQLC: number;
    U_STK: number;
    U_QLC: number;
    U_QLC2: number;
    U_QLC3: number;
    U_TotalPrice: number;
    U_MK_THB: number;
    U_SalesPrice: number;
}

export interface TermCalcResults {
    OP1: number;
    OP1_THB: number;
    OP2_THB: number;
    DIM_WEIGHT: number;
    SHP_WEIGHT: number;
    INS: number;
    FR_QTEC: number;
    FR_ZONE: number;
    CIF: number;
    CIF_ZONE: number;
    DT: number;
    DT_FR: number;
    DT_ZONE: number;
    ET: number;
    MT: number;
    PRE_QLC: number;
    STK: number;
    QLC: number;
    QLC2: number;
    QLC3: number;
    TOTAL_PRICE: number;
    MK_THB: number;
    SALES_PRICE: number;
}

export interface TermStageStatus {
    OP1: boolean;
    FR: boolean;
    INS: boolean;
    CIF: boolean;
    DT: boolean;
    ET: boolean;
    MT: boolean;
    TERM: boolean;
    UOM: boolean;
    QLC: boolean;
}